move definitions
