+++ library/common/include/BUILD
@@ -6,6 +6,7 @@ envoy_package()
 
 envoy_cc_library(
     name = "c_types_interface",
+    srcs = ["c_types.cc"],
     hdrs = ["c_types.h"],
     repository = "@envoy",
 )
